backend/requirements.txt
beautifulsoup4
requests
psycopg2-binary
uvicorn
fastapi